requirements.txt
firebase-admin
google-cloud-firestore
requests
flask
Flask-Cors
python-dotenv
mailjet-rest
